creando trash content

+++ app/Http/Controllers/ProjectController.php
@@ -6,6 +6,7 @@ use App\Category;
 use App\Events\ProjectSaved;
 use http\Env\Request;
 use App\Models\Project;
+use Illuminate\Auth\Access\Gate;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\SaveProjectRequest;
 
@@ -19,6 +20,7 @@ class ProjectController extends Controller
     public function index()
     {
         return view('projects.index', [
+            'newProject' => new Project,
             'projects' => Project::latest()->paginate(),
         ]);
     }
@@ -32,9 +34,11 @@ class ProjectController extends Controller
 
     public function create()
     {
+        $this->authorize('create', $project = new Project());
+
         return view('projects.create', [
-            'project' => new Project,
-            'categories' => Category::pluck('name', 'id')
+            'project' => $project,
+            'categories' => Category::pluck('name', 'id'),
         ]);
     }
 
@@ -42,26 +46,31 @@ class ProjectController extends Controller
     {
         $project = new Project($request->validated());
 
+        $this->authorize('create', $project);
+
         $project->image = $request->file('image')->store('images');
 
         $project->save();
 
         ProjectSaved::dispatch($project);
 
-
         return redirect()->route('projects.index')->with('status', 'El proyecto fue creado con éxito');
     }
 
     public function edit(Project $project)
     {
+        $this->authorize('update', $project);
+
         return view('projects.edit', [
             'project' => $project,
-            'categories' => Category::pluck('name', 'id')
+            'categories' => Category::pluck('name', 'id'),
         ]);
     }
 
     public function update(Project $project, SaveProjectRequest $request)
     {
+        $this->authorize('updtade', $project);
+
         if ($request->hasFile('image')) {
             Storage::delete($project->image);
 
@@ -81,6 +90,8 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('delete    ', $project);
+
         Storage::delete($project->image);
 
         $project->delete();
